Separate and group utility classes and files
