+++ Sources/SwiftUICharts/BarChart/BarChartView.swift
@@ -135,14 +135,3 @@ public struct BarChartView : View {
         return self.data.points[index]
     }
 }
-
-#if DEBUG
-struct ChartView_Previews : PreviewProvider {
-    static var previews: some View {
-        BarChartView(data: TestData.values ,
-                     title: "Model 3 sales",
-                     legend: "Quarterly",
-                     valueSpecifier: "%.0f")
-    }
-}
-#endif
